Added current engine files
Added the current engine files to a new visual studio proejct. Going to begin working on this later on and will show Borna where to pull from and run the tests. If I burn through the testing in general, I'll try and write some quick continuous integration files to see where it can go from there